+++ tictactoe.html
@@ -5,10 +5,10 @@
 		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<title>Tic Tac Toe</title>
-		<script src="script.js" defer></script>
+		<script src="js/script.js" defer type="module"></script>
 		<link rel="stylesheet" href="style.css" />
 	</head>
-	<body onload="openNav()">
+	<body>
 		<div id="select_gamemode" class="overlay">
 			<div class="overlay_content">
 				<h1>SELECT GAMEMODE</h1>
